add tests to main file
